pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "security-ai-eval-lab"
version = "0.1.0"
description = "Evaluation lab for LLM-assisted security investigation tasks."
requires-python = ">=3.11"
dependencies = [
    # Pulled from GitHub Packages. Install before `pip install -e .` with:
    #   pip install --extra-index-url "https://__token__:${REGISTRY_TOKEN}@pypi.pkg.github.com/ngallodev-software/" ai-reliability-fw==0.1.0
    "ai-reliability-fw==0.1.0",
    "alembic==1.18.4",
    "openai>=1.0.0",
    "asyncpg==0.30.0",
    "jsonschema==4.26.0",
    "python-dotenv==1.2.2",
    "SQLAlchemy==2.0.48",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "db*", "evaluation*", "signals*", "llm*"]
namespaces = true
